Figuring out the framework
