+++ src/main/java/ca/mcmaster/se2aa4/island/team45/drone/DirectionManager.java
@@ -1,4 +1,4 @@
-package ca.mcmaster.se2aa4.island.team45;
+package ca.mcmaster.se2aa4.island.team45.drone;
 
 public class DirectionManager {
     private String direction;
